Added audio noise generation.
This is basically cribbed straight from MDN's tutorial for audiobuffer.
I've used audiobuffer before, to capture input directly from the microphone
and store it for an analyzer to, uh, analyze, but this was just too easy.

MDN's example just generates white noise, which is exactly what I needed.
The only modifications were to add a gain node and to randomly adjust the
value of the gain within 0.3 and 1.0 every 60ms.  This gives the static a
sort of non-random randomness that makes it a little more like television
static and a little less like pure random noise.